Update bonsai-gym to support newer version of Gym
Newer versions of ``gym`` broke compatibility with ``bonsai-gym``.
This commit refactors the package to support the latest version of the
``gym``.
It also remove all dependencies from the deprecated ``bonsai-common``.
When reimplementing the functionalities of bonsai-common into the ``bonsai-gym``
we broke the API for ``GymSimulator3``. Now it does not require to provide a
``BonsaiClientConfig`` instance at initialization.